Convert VFSFSAL_readdir to use the vfsfsal_get_dirent API
This also changes the way that startposition and end_position are used.
Previously the record offset of the last record was used as an lseek
pointer to start the next call.  Note that this causes that record to
be examined twice (assuming the lseek worked).  However, that was masked
because everything is put into the cache by cache_inode_readdir_populate,
so the duplicate gets eliminated.

However, BSD systems don't define a record offset for each dirent, so
it is impossible to lseek into the middle of a batch of dirents and get
back to where you think you might want to be.  In short, you can only
lseek to the boundary between sucessive getdents() calls.

To handle this, the start position is now an lseek pointer and a number of
records to skip over on the subsequent call.  Recall that cache_inode_readdir
has a buffer of 2048 directory entries.  The FSAL readdir will call getdents
and get chunks of dirents in a loop.  Eventually the caller's buffer fills
up, and only part of the last chunk read by getdents will have been saved
in the caller's buffer.  That number becomes the number of records to skip
on the next call to FSAL_readdir, which will seek back and read the last
chunk a second time, and skip over stuff it has already given to it caller
on the previous iteration.

The reason for this complex dance is that the cache_inode_readdir has this
big temporary buffer that it is going to throw away anyway.  A different API
would allow the FSAL to allocate memory for a batch of dirents, and then
cache_inode_readdir_populate would free that after it stuffs things into
its cache.

Signed-off-by: Brent Welch <welch@panasas.com>