Fix VFSFSAL_DigestHandle to create better 64-bit digests for PanFS handles
This is based on the new handle API.

The FILEID3 and FILEID4 cases are collapsed together because the
digests are the same size (64 bits).

Add VFSFSAL_handle_printf that returns char *

This change introduces a #ifdef LINUX / #ifdef FREEBSD branch that should
be an API instead.  However, this code is going to be replaced with a
different way to compute state IDs, so hopefully this just goes away.

Signed-off-by: Brent Welch <welch@panasas.com>